backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.1
minio==7.2.7
python-multipart==0.0.9
